Se separo logica de vista
